+++ package.json
@@ -13,6 +13,7 @@
     "@nuxtjs/google-fonts": "^3.2.0",
     "flowbite": "^2.3.0",
     "nuxt": "^3.11.2",
+    "vee-validate": "^4.13.1",
     "vue": "^3.4.27",
     "vue-router": "^4.3.2"
   },
